+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Add Solr transformation to add taxonomy term ancestors to the list of values
+  (requires Solr module 0.13.0)
+
 ## [0.6.0] - 2023-12-05
 
 - In admin "show-details" sidebar, make the linked resources count clickable